requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
pika>=1.3.2
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── ML / Classifier Dependencies ──
transformers>=4.40.0
torch>=2.2.0
datasets>=2.19.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
onnxruntime>=1.17.0
accelerate>=0.28.0

# ── Agentic Workflow ──
langgraph>=0.2.0


